# apps/server/openapi.json
{"openapi":"3.1.0","paths":{"/api":{"get":{"operationId":"AppController_getData","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/hello":{"get":{"operationId":"AppController_hello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/v1/accounts/sms-codes":{"post":{"description":"Generates a 6-digit code, stores in Redis with TTL, dispatches via SmsGateway. Rate-limited per phone (60s + 24h 10) and per IP (24h 50).","operationId":"AccountSmsCodeController_request","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestSmsCodeRequest"}}}},"responses":{"200":{"description":"Code dispatched — client should start ttlSec countdown","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestSmsCodeResponse"}}}},"400":{"description":"Validation failure (invalid phone format)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S07 #1-3)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"summary":"Request an SMS verification code","tags":["accounts"]}},"/api/v1/accounts/phone-sms-auth":{"post":{"description":"Verifies SMS code; auto-registers unregistered phones (US2); returns identical shape across all paths for byte-level anti-enumeration (US3).","operationId":"AccountPhoneSmsAuthController_auth","parameters":[{"name":"x-device-id","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneSmsAuthRequest"}}}},"responses":{"200":{"description":"Auth success — tokens issued","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneSmsAuthResponse"}}}},"400":{"description":"Validation failure (invalid phone format / code format)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Anti-enumeration auth failure (wrong code / expired / etc.)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"403":{"description":"Account in 30-day freeze period (FR-S11; FROZEN disclosure)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Auth attempts locked (FR-S07 #4; 5 failures → 30-min lock)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"summary":"Authenticate via phone + SMS code","tags":["accounts"]}},"/api/v1/accounts/refresh-token":{"post":{"description":"Atomically revokes the presented refresh token and issues a fresh access + refresh pair (single-use). Anti-enumeration: all failures fold to 401.","operationId":"AccountTokenController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}}},"responses":{"200":{"description":"Rotated — new tokens issued","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneSmsAuthResponse"}}}},"400":{"description":"Validation failure (empty / missing refreshToken)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Anti-enumeration failure (invalid / expired / revoked / account not eligible)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S14: per-IP 100/60s, per-token 5/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"summary":"Rotate a refresh token","tags":["accounts"]}},"/api/v1/accounts/logout-all":{"post":{"description":"Revokes every active refresh token for the bearer-authenticated account (current device included). Idempotent → 204.","operationId":"AccountTokenController_logoutAll","parameters":[],"responses":{"204":{"description":"All sessions revoked (idempotent)"},"401":{"description":"Missing / invalid / expired access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S14: per-IP 50/60s, per-account 5/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"bearer":[]}],"summary":"Log out from all devices","tags":["accounts"]}},"/api/v1/accounts/me/deletion-codes":{"post":{"description":"Issues a DELETE_ACCOUNT SMS code for the bearer-authenticated account (FR-S01/S02). Non-ACTIVE accounts fold to 401 (anti-enumeration). 204 on dispatch.","operationId":"AccountDeletionController_sendDeletionCodeForMe","parameters":[],"responses":{"204":{"description":"Code dispatched (no body)"},"401":{"description":"Invalid token or account not ACTIVE — reason not disclosed (anti-enumeration)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S18: per-account 1/60s, per-IP 5/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"503":{"description":"SMS dispatch failed (FR-S21)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"bearer":[]}],"summary":"Send account-deletion verification code (EP1)","tags":["account-deletion"]}},"/api/v1/accounts/me/deletion":{"post":{"description":"Validates the DELETE_ACCOUNT code then atomically freezes the account (15-day grace), revokes all refresh tokens, and emits the deletion-requested event (FR-S03/S04/S06). Any code failure folds to 401 INVALID_DELETION_CODE. 204 on freeze.","operationId":"AccountDeletionController_submitDeletionForMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteAccountRequest"}}}},"responses":{"204":{"description":"Account frozen (no body)"},"400":{"description":"Missing / non-6-digit code — FORM_VALIDATION (distinct from credential path)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Code not found / hash mismatch / expired / used — folded to INVALID_DELETION_CODE (anti-enumeration)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S18: per-account 5/60s, per-IP 10/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"bearer":[]}],"summary":"Submit account-deletion code → freeze account (EP2)","tags":["account-deletion"]}},"/api/v1/auth/cancel-deletion/sms-codes":{"post":{"description":"Public endpoint: accepts a phone number; only FROZEN-in-grace accounts receive a CANCEL_DELETION SMS code. The 4 ineligible classes (unregistered / ACTIVE / ANONYMIZED / grace-expired) silently return 200 with a dummy timing pad — no code written, no SMS — so eligible/ineligible are byte- and timing-indistinguishable (FR-S07 anti-enumeration).","operationId":"CancelDeletionController_sendCancelCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCancelCodeRequest"}}}},"responses":{"200":{"description":"Accepted — code sent iff eligible (response indistinguishable from ineligible)"},"422":{"description":"Phone not E.164 +86 CN mobile — INVALID_PHONE_FORMAT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S18: per-phone-hash 1/60s, per-IP 5/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"503":{"description":"SMS dispatch failed on the eligible path (FR-S21)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"summary":"Send cancel-deletion verification code (EP3, public)","tags":["account-deletion"]}},"/api/v1/auth/cancel-deletion":{"post":{"description":"Public endpoint: FROZEN-in-grace account + valid CANCEL_DELETION code atomically unfreezes (FROZEN→ACTIVE), issues a fresh login session, and emits the deletion-cancelled event (FR-S09/S10/S12). All 5 failure classes (unregistered / ACTIVE / ANONYMIZED / grace-expired / invalid code) fold to byte-identical 401 INVALID_CREDENTIALS with a timing pad (FR-S11 anti-enumeration). 200 + LoginResponse on success.","operationId":"CancelDeletionController_cancelDeletion","parameters":[{"name":"x-device-id","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelDeletionRequest"}}}},"responses":{"200":{"description":"Unfrozen — fresh login tokens","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneSmsAuthResponse"}}}},"400":{"description":"Missing field / non-6-digit code — FORM_VALIDATION","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Account-state / code failure folded to INVALID_CREDENTIALS (anti-enumeration)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"422":{"description":"Phone not E.164 +86 CN mobile — INVALID_PHONE_FORMAT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-S18: per-phone-hash 5/60s, per-IP 10/60s)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"summary":"Submit cancel-deletion code → unfreeze + re-login (EP4, public)","tags":["account-deletion"]}},"/api/v1/accounts/me":{"get":{"description":"Returns account profile for the bearer-authenticated user. Phone is E.164 raw string; displayName is null for new users (FR-007).","operationId":"AccountProfileController_getProfile","parameters":[],"responses":{"200":{"description":"Profile retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountProfileResponse"}}}},"401":{"description":"Missing / invalid / expired token, or account not ACTIVE (FR-002, FR-009) — reason not disclosed (anti-enumeration)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-008: 60 requests per 60s per account)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"bearer":[]}],"summary":"Get authenticated account profile","tags":["accounts"]},"patch":{"description":"Sets a new display name for the bearer-authenticated user. Validates FR-005 rules (1-32 Unicode code points, no forbidden chars). Returns updated profile.","operationId":"AccountProfileController_updateDisplayName","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDisplayNameRequest"}}}},"responses":{"200":{"description":"Display name updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountProfileResponse"}}}},"400":{"description":"Invalid display name (violates FR-005 rules)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"401":{"description":"Missing / invalid / expired token, or account not ACTIVE (FR-004, FR-009) — reason not disclosed (anti-enumeration)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"429":{"description":"Rate limit exceeded (FR-008: 10 requests per 60s per account)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}}},"security":[{"bearer":[]}],"summary":"Update authenticated account display name","tags":["accounts"]}},"/healthz/live":{"get":{"operationId":"HealthController_live","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/healthz/ready":{"get":{"operationId":"HealthController_ready","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/metrics":{"get":{"operationId":"PrometheusController_index","parameters":[],"responses":{"200":{"description":""}},"tags":["Prometheus"]}}},"info":{"title":"no-vain-years API","description":"no-vain-years backend HTTP API. Generated from NestJS controllers; consumed by packages/api-client for cross-app TS types.","version":"1.0","contact":{}},"tags":[{"name":"accounts","description":"Account lifecycle: SMS code request, phone-SMS auth"}],"servers":[],"components":{"schemas":{"RequestSmsCodeRequest":{"type":"object","properties":{"phone":{"type":"string","description":"E.164 +86 CN mobile phone number","example":"+8613800138000","pattern":"^\\+861[3-9]\\d{9}$"}},"required":["phone"]},"RequestSmsCodeResponse":{"type":"object","properties":{"ttlSec":{"type":"number","description":"SMS code TTL in seconds (client countdown / resend gating)","example":60,"minimum":1}},"required":["ttlSec"]},"ProblemDetailResponse":{"type":"object","properties":{"type":{"type":"string","description":"RFC 9457 problem type URI; \"about:blank\" if generic","example":"about:blank"},"title":{"type":"string","description":"Short human-readable summary of the problem type","example":"Forbidden"},"status":{"type":"number","description":"HTTP status code","example":403},"detail":{"type":"string","description":"Human-readable explanation specific to this occurrence","example":"Account is in 30-day freeze period"},"instance":{"type":"string","description":"URI reference identifying the specific occurrence","example":"/api/v1/accounts/phone-sms-auth"},"code":{"type":"string","description":"Machine-readable error code (UPPER_SNAKE_CASE)","example":"ACCOUNT_IN_FREEZE_PERIOD"},"traceId":{"type":"string","description":"Request trace id (CLS-managed; mirrors x-trace-id response header)","example":"0e6a4d6e-...-2c8f"},"freezeUntil":{"type":"string","description":"ISO 8601 timestamp — for ACCOUNT_IN_FREEZE_PERIOD","example":"2026-06-20T10:00:00.000Z"},"retryAfterSeconds":{"type":"number","description":"Seconds until retry permitted — for AUTH_ATTEMPT_LOCKED / RATE_LIMIT_EXCEEDED","example":1800},"invalidAttributes":{"type":"array","description":"Per-attribute validation issues — for FORM_VALIDATION; drives client form.setError()","items":{"type":"object","properties":{"field":{"type":"string","example":"displayName"},"messages":{"type":"array","items":{"type":"string","example":"1-32 chars required"}}}}}},"required":["type","title","status"]},"PhoneSmsAuthRequest":{"type":"object","properties":{"phone":{"type":"string","description":"E.164 +86 CN mobile phone number","example":"+8613800138000","pattern":"^\\+861[3-9]\\d{9}$"},"code":{"type":"string","description":"6-digit SMS verification code","example":"123456","pattern":"^\\d{6}$"}},"required":["phone","code"]},"PhoneSmsAuthResponse":{"type":"object","properties":{"accountId":{"type":"string","description":"Account ID (serialized as string to remain JSON-safe vs BigInt; matches JWT sub claim)","example":"1234567890"},"accessToken":{"type":"string","description":"JWT access token (short-lived)","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"refreshToken":{"type":"string","description":"JWT refresh token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["accountId","accessToken","refreshToken"]},"RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Opaque refresh token issued at login / previous rotation","example":"dGhpcy1pcy1hLXJlZnJlc2gtdG9rZW4tc2FtcGxl"}},"required":["refreshToken"]},"DeleteAccountRequest":{"type":"object","properties":{"code":{"type":"string","description":"6-digit account-deletion verification code","example":"123456","pattern":"^\\d{6}$"}},"required":["code"]},"SendCancelCodeRequest":{"type":"object","properties":{"phone":{"type":"string","description":"E.164 +86 CN mobile phone number","example":"+8613800138000","pattern":"^\\+861[3-9]\\d{9}$"}},"required":["phone"]},"CancelDeletionRequest":{"type":"object","properties":{"phone":{"type":"string","description":"E.164 +86 CN mobile phone number","example":"+8613800138000","pattern":"^\\+861[3-9]\\d{9}$"},"code":{"type":"string","description":"6-digit cancel-deletion verification code","example":"123456","pattern":"^\\d{6}$"}},"required":["phone","code"]},"AccountProfileResponse":{"type":"object","properties":{"accountId":{"type":"string","description":"Account ID (serialized as string for JSON-safety vs BigInt; matches JWT sub claim)","example":"1234567890"},"phone":{"type":"string","description":"E.164 phone number; mask handled by client (per apps/mobile/lib/format/phone.ts)","example":"+8613800138000"},"displayName":{"type":"string","description":"Display name; null for new users before first update (FR-007)","example":"Alice","nullable":true},"status":{"type":"string","description":"Account status","enum":["ACTIVE","FROZEN","ANONYMIZED"],"example":"ACTIVE"},"createdAt":{"type":"string","description":"Account creation timestamp (ISO 8601)","example":"2026-01-01T00:00:00.000Z","format":"date-time"}},"required":["accountId","phone","displayName","status","createdAt"]},"UpdateDisplayNameRequest":{"type":"object","properties":{"displayName":{"type":"string","description":"New display name; 1-32 Unicode code points after trim (FR-005)","example":"Alice","maxLength":32}},"required":["displayName"]}}}}